r_matroska: avoid deriving any ref for SimpleBlock
The only reason a bref or fref has been faked for a SimpleBlock is
that we hasn't been checking whether it is a keyframe in the right
way. It's a bad idea anyway as the referencing models differ among
video codecs, let alone files that do not follow the rules.