packet: remove ambiguous is_{p,b}_frame()
It appears to be a bad idea to equate the ref(s) a packet has with
its frame type, as the definition of the latter differ among video
codecs.

In the only usage of is_p_frame(), it is used together with
is_key_frame() to determine how many ReferenceBlock should be
considered when calculating a byte count, but it's actually more
straight-forward to determine with has_bref() and has_fref().
(It is unclear whether the formula is correct for SimpleBlock
anyway, but at least it will one step closer to being so, given
the fact that a SimpleBlock does not have any ReferenceBlock,
regardless of its frame type.)

is_b_frame() is not used anywhere.